scripts/ai_review/requirements.txt
python-dotenv>=1.0.0
openai>=2.6.1
GitPython>=3.1.0

# Testing dependencies
pytest>=7.4.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Development dependencies (code quality)
black>=23.0.0
ruff>=0.1.0
